+++ public/static/home/js/chapter.live.player.js
@@ -6,41 +6,53 @@ layui.use(['jquery', 'helper'], function () {
     var interval = null;
     var intervalTime = 15000;
     var userId = window.user.id;
+    var requestId = helper.getRequestId();
     var planId = $('input[name="chapter.plan_id"]').val();
     var learningUrl = $('input[name="chapter.learning_url"]').val();
     var playUrls = JSON.parse($('input[name="chapter.play_urls"]').val());
-    var requestId = helper.getRequestId();
 
-    var options = {
-        live: true,
-        autoplay: true,
-        width: 760,
-        height: 428,
-    };
+    var formats = ['flv'];
 
-    var formats = ['m3u8'];
-    var rates = ['od', 'hd', 'sd'];
+    var rates = [
+        {name: 'od', label: '原画'},
+        {name: 'hd', label: '高清'},
+        {name: 'sd', label: '标清'},
+    ];
+
+    var quality = [];
 
     $.each(formats, function (i, format) {
         $.each(rates, function (k, rate) {
-            if (playUrls.hasOwnProperty(format) && playUrls[format].hasOwnProperty(rate)) {
-                var key = k === 0 ? format : format + '_' + rate;
-                options[key] = playUrls[format][rate];
+            if (playUrls.hasOwnProperty(format) && playUrls[format].hasOwnProperty(rate.name)) {
+                quality[k] = {
+                    name: rate.label,
+                    url: playUrls[format][rate.name],
+                    type: 'flv',
+                };
             }
         });
     });
 
-    options.listener = function (msg) {
-        if (msg.type === 'play') {
-            start();
-        } else if (msg.type === 'pause') {
-            stop();
-        } else if (msg.type === 'ended') {
-            stop();
+    var player = new DPlayer({
+        container: document.getElementById('player'),
+        live: true,
+        video: {
+            quality: quality,
+            defaultQuality: 0,
         }
-    };
+    });
 
-    var player = new TcPlayer('player', options);
+    player.on('play', function () {
+        start();
+    });
+
+    player.on('pause', function () {
+        stop();
+    });
+
+    player.on('ended', function () {
+        stop();
+    });
 
     function start() {
         if (interval != null) {